[ADD] app: Add new status bar states
The file `internal/app/update.go` was modified to add new states for the status bar. Specifically, the `getDefaultStatusBarContent` function was updated to return custom status bar content for the following states: "Creating a new Rune", "Editing Rune", "Executing Rune", and "Creating a new Loeg". These changes directly support the addition of new states for the status bar as stated in the title.