[FIX] app: Cancel editing rune bug fix
The `internal/app/update.go` file was modified to fix an issue where canceling the editing of a rune would incorrectly trigger a spellbook update. Specifically, the `updateEditingRune` function was updated to reset the `StatusBar` content and level when editing a rune is canceled, and to not call `m.getSpellbookContentCmd` in this case.